Add show-lists.js as command
The show-cards command requires the list name to be passed in, as well
as the board name.  However, we have no way of knowing the lists that
are a part of a board.  Added a command that shows all of the lists, or
all of the lists in a specified board